requirements.txt
# AWS
boto3==1.34.0

# GCP
google-cloud-billing==1.13.0
google-cloud-resource-manager==1.12.0

# Time-series DB
influxdb-client==1.40.0

# RAG / Embeddings
sentence-transformers==2.6.1
pinecone-client==3.1.0

# LLM
anthropic==0.25.0

# GitHub integration
PyGithub==2.3.0

# Scheduling
APScheduler==3.10.4

# Utilities
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0
slack-sdk==3.27.1
beautifulsoup4==4.12.3

# Testing
pytest==8.0.0
pytest-mock==3.12.0
